requirements.txt
PyYAML
gitpython
schema
